Added validation

+++ src/main/java/server/Dot.java
@@ -5,12 +5,12 @@ import java.time.LocalDate;
 public class Dot {
     private int x;
     private double y;
-    private double r;
+    private int r;
     private LocalDate currentTime;
     private long scriptTime;
     private String result;
 
-    public Dot(int x, double y, double r, LocalDate currentTime, long scriptTime, String result){
+    public Dot(int x, double y, int r, LocalDate currentTime, long scriptTime, String result){
         this.x = x;
         this.y = y;
         this.r = r;
@@ -27,7 +27,7 @@ public class Dot {
         return y;
     }
 
-    public double getR() {
+    public int getR() {
         return r;
     }
